requirements.txt
python-dotenv==1.1.1
requests==2.32.3
rich==13.7.1
apscheduler==3.10.4
pytz==2024.1
slack-sdk==3.27.1
